增加ActionHandlerBuilder
